@@ -30,29 +30,21 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>8</java.version>
+        <java.version>11</java.version>
         <maven.test.skip>false</maven.test.skip>
         <apt-maven-plugin.version>1.1.3</apt-maven-plugin.version>
-        <dependency-check-maven.version>5.2.0</dependency-check-maven.version>
-        <jacoco-maven-plugin.version>0.8.4</jacoco-maven-plugin.version>
-        <commons-beanutils.version>1.9.3</commons-beanutils.version>
-        <commons-text.version>1.7</commons-text.version>
-        <spring-cloud-sleuth.version>2.1.2.RELEASE</spring-cloud-sleuth.version>
+        <dependency-check-maven.version>5.2.4</dependency-check-maven.version>
+        <jacoco-maven-plugin.version>0.8.5</jacoco-maven-plugin.version>
+        <commons-beanutils.version>1.9.4</commons-beanutils.version>
+        <commons-text.version>1.8</commons-text.version>
+        <spring-cloud.version>Hoxton.RC2</spring-cloud.version>
     </properties>
 
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <groupId>org.springframework.cloud</groupId>
-                <artifactId>spring-cloud-sleuth</artifactId>
-                <version>${spring-cloud-sleuth.version}</version>
-                <type>pom</type>
-                <scope>import</scope>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
-
     <dependencies>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-web</artifactId>
+        </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-actuator</artifactId>
