@@ -15,28 +15,21 @@
  */
 package io.undertree.symptom.repositories;
 
-import javax.validation.ConstraintViolationException;
-
 import io.undertree.symptom.domain.GivenName;
 import io.undertree.symptom.domain.Patient;
 import io.undertree.symptom.domain.TestPatientBuilder;
 import io.undertree.symptom.exceptions.NotFoundException;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.ExpectedException;
-import org.junit.runner.RunWith;
-
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.dao.InvalidDataAccessApiUsageException;
-import org.springframework.test.context.junit4.SpringRunner;
 
+import javax.validation.ConstraintViolationException;
 import java.time.LocalDate;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.hamcrest.Matchers.allOf;
-import static org.hamcrest.Matchers.containsString;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 // refer to:
 // https://joel-costigliola.github.io/assertj/
@@ -44,100 +37,133 @@ import static org.hamcrest.Matchers.containsString;
 // https://spring.io/blog/2016/04/15/testing-improvements-in-spring-boot-1-4
 // http://docs.spring.io/spring/docs/current/spring-framework-reference/htmlsingle/#__sql
 
-@RunWith(SpringRunner.class)
 @DataJpaTest
 public class PatientRepositoryTests {
 
-	@Rule
-	public ExpectedException thrown = ExpectedException.none();
-	@Autowired
-	PatientRepository patientRepository;
-	@Autowired
-	private TestEntityManager entityManager;
+    @Autowired
+    PatientRepository patientRepository;
 
-	@Test
-	public void test_PatientRepository_findByPatientId_ExpectExists() throws Exception {
-		Patient patient = entityManager.persistFlushFind(new TestPatientBuilder().build());
-		Patient aPatient = patientRepository.findByPatientId(patient.getPatientId()).orElseThrow(NotFoundException::new);
-		assertThat(aPatient).isEqualTo(patient);
-	}
+    @Autowired
+    private TestEntityManager entityManager;
 
-	@Test
-	public void test_PatientRepository_SaveWithNull_ExpectException() throws Exception {
-		thrown.expect(InvalidDataAccessApiUsageException.class);
-		patientRepository.saveAndFlush(null);
-	}
+    @Test
+    public void test_PatientRepository_findByPatientId_ExpectExists() throws Exception {
+        Patient patient = entityManager.persistFlushFind(new TestPatientBuilder().build());
+        Patient aPatient = patientRepository.findByPatientId(patient.getPatientId()).orElseThrow(NotFoundException::new);
+        assertThat(aPatient).isEqualTo(patient);
+    }
 
-	@Test
-	public void test_PatientRepository_SaveWithEmpty_ExpectException() throws Exception {
-		thrown.expect(ConstraintViolationException.class);
-		thrown.expectMessage(containsString("'must not be blank'"));
-		patientRepository.saveAndFlush(new Patient());
-	}
+    @Test
+    public void test_PatientRepository_SaveWithNull_ExpectException() throws Exception {
+        assertThrows(InvalidDataAccessApiUsageException.class, () -> {
+            patientRepository.saveAndFlush(null);
+        });
+    }
 
-	@Test
-	public void test_PatientRepository_SaveWithEmptyGivenName_ExpectException() throws Exception {
-		thrown.expect(ConstraintViolationException.class);
-		thrown.expectMessage(allOf(containsString("givenName.givenName"),
-				containsString("'The given name should only contain alphanumeric values.'")));
-		thrown.expectMessage(allOf(containsString("givenName.givenName"),
-				containsString("'size must be between 2 and 50'")));
-		patientRepository.saveAndFlush(new TestPatientBuilder().withGivenName(new GivenName("")).build());
-	}
+    @Test
+    public void test_PatientRepository_SaveWithEmpty_ExpectException() throws Exception {
+        ConstraintViolationException exception =
+				assertThrows(ConstraintViolationException.class,
+						() -> patientRepository.saveAndFlush(new Patient()));
 
-	@Test
-	public void test_PatientRepository_SaveWithEmptyFamilyName_ExpectException() throws Exception {
-		thrown.expect(ConstraintViolationException.class);
-		thrown.expectMessage(allOf(containsString("familyName"), containsString("'must not be blank'")));
-		patientRepository.saveAndFlush(new TestPatientBuilder().withFamilyName("").build());
-	}
+        assertThat(exception).hasMessageContaining("'must not be blank'");
+    }
 
-	@Test
-	public void test_PatientRepository_SaveWithShortGivenName_ExpectException() throws Exception {
-		thrown.expect(ConstraintViolationException.class);
-		thrown.expectMessage(
-				allOf(containsString("givenName"), containsString("'size must be between 2 and")));
-		patientRepository.saveAndFlush(new TestPatientBuilder().withGivenName(new GivenName("A")).build());
-	}
+    @Test
+    public void test_PatientRepository_SaveWithEmptyGivenName_ExpectException() throws Exception {
+        ConstraintViolationException exception =
+				assertThrows(ConstraintViolationException.class,
+						() -> patientRepository.saveAndFlush(new TestPatientBuilder().withGivenName(new GivenName("")).build()));
 
-	@Test
-	public void test_PatientRepository_SaveWithShortFamilyName_ExpectException() throws Exception {
-		thrown.expect(ConstraintViolationException.class);
-		thrown.expectMessage(
-				allOf(containsString("familyName"), containsString("'size must be between 2 and")));
-		patientRepository.saveAndFlush(new TestPatientBuilder().withFamilyName("Z").build());
-	}
+        assertThat(exception)
+                .hasMessageContaining("givenName.givenName")
+                .hasMessageContaining("'The given name should only contain alphanumeric values.'")
+                .hasMessageContaining("'size must be between 2 and 50'");
+    }
 
-	@Test
-	public void test_PatientRepository_SaveWithInvalidEmail_ExpectException() throws Exception {
-		thrown.expect(ConstraintViolationException.class);
-		thrown.expectMessage(
-				allOf(containsString("email"), containsString("'must be a well-formed email address'")));
-		patientRepository.saveAndFlush(new TestPatientBuilder().withEmail("baz").build());
-	}
+    @Test
+    public void test_PatientRepository_SaveWithEmptyFamilyName_ExpectException() throws Exception {
 
-	@Test
-	public void test_PatientRepository_SaveWithLessThanMinHeight_ExpectException() throws Exception {
-		thrown.expect(ConstraintViolationException.class);
-		thrown.expectMessage(
-				allOf(containsString("height"), containsString("'must be greater than or equal to 0'")));
-		patientRepository.saveAndFlush(new TestPatientBuilder().withHeight((short) -1).build());
-	}
+        ConstraintViolationException exception =
+				assertThrows(ConstraintViolationException.class,
+						() -> patientRepository.saveAndFlush(new TestPatientBuilder().withFamilyName("").build()));
 
-	@Test
-	public void test_PatientRepository_SaveWithLessThanMinWeight_ExpectException() throws Exception {
-		thrown.expect(ConstraintViolationException.class);
-		thrown.expectMessage(
-				allOf(containsString("weight"), containsString("'must be greater than or equal to 0'")));
-		patientRepository.saveAndFlush(new TestPatientBuilder().withWeight((short) -1).build());
-	}
+        assertThat(exception)
+                .hasMessageContaining("familyName")
+                .hasMessageContaining("'must not be blank'");
+    }
 
-	@Test
-	public void test_PatientRepository_SaveWithFutureBirthDayInFuture_ExpectException() throws Exception {
-		thrown.expect(ConstraintViolationException.class);
-		thrown.expectMessage(
-				allOf(containsString("birthDate"), containsString("'must be a past date'")));
-		patientRepository.saveAndFlush(new TestPatientBuilder()
-				.withBirthDate(LocalDate.now().plusDays(1)).build());
-	}
+    @Test
+    public void test_PatientRepository_SaveWithShortGivenName_ExpectException() throws Exception {
+
+    	ConstraintViolationException exception =
+				assertThrows(ConstraintViolationException.class,
+						() -> patientRepository.saveAndFlush(new TestPatientBuilder().withGivenName(new GivenName("A")).build()));
+
+        assertThat(exception)
+                .hasMessageContaining("givenName")
+                .hasMessageContaining("'size must be between 2 and");
+    }
+
+    @Test
+    public void test_PatientRepository_SaveWithShortFamilyName_ExpectException() throws Exception {
+
+        ConstraintViolationException exception =
+				assertThrows(ConstraintViolationException.class,
+						() -> patientRepository.saveAndFlush(new TestPatientBuilder().withFamilyName("Z").build()));
+
+        assertThat(exception)
+                .hasMessageContaining("familyName")
+                .hasMessageContaining("'size must be between 2 and");
+    }
+
+
+    @Test
+    public void test_PatientRepository_SaveWithInvalidEmail_ExpectException() throws Exception {
+
+        ConstraintViolationException exception =
+				assertThrows(ConstraintViolationException.class,
+						() -> patientRepository.saveAndFlush(new TestPatientBuilder().withEmail("baz").build()));
+
+        assertThat(exception)
+                .hasMessageContaining("email")
+                .hasMessageContaining("'must be a well-formed email address'");
+    }
+
+
+    @Test
+    public void test_PatientRepository_SaveWithLessThanMinHeight_ExpectException() throws Exception {
+
+        ConstraintViolationException exception =
+				assertThrows(ConstraintViolationException.class,
+						() -> patientRepository.saveAndFlush(new TestPatientBuilder().withHeight((short) -1).build()));
+
+        assertThat(exception)
+                .hasMessageContaining("height")
+                .hasMessageContaining("'must be greater than or equal to 0'");
+    }
+
+
+    @Test
+    public void test_PatientRepository_SaveWithLessThanMinWeight_ExpectException() throws Exception {
+
+        ConstraintViolationException exception =
+				assertThrows(ConstraintViolationException.class,
+						() -> patientRepository.saveAndFlush(new TestPatientBuilder().withWeight((short) -1).build()));
+
+        assertThat(exception)
+                .hasMessageContaining("weight")
+                .hasMessageContaining("'must be greater than or equal to 0'");
+    }
+
+    @Test
+    public void test_PatientRepository_SaveWithFutureBirthDayInFuture_ExpectException() throws Exception {
+        ConstraintViolationException exception =
+				assertThrows(ConstraintViolationException.class,
+						() -> patientRepository.saveAndFlush(new TestPatientBuilder().withBirthDate(LocalDate.now().plusDays(1)).build()));
+
+        assertThat(exception)
+                .hasMessageContaining("birthDate")
+                .hasMessageContaining("'must be a past date'");
+    }
 }
