@@ -15,29 +15,25 @@
  */
 package io.undertree.symptom.controllers;
 
+import io.undertree.symptom.domain.GivenName;
+import io.undertree.symptom.domain.TestPatientBuilder;
+import io.undertree.symptom.repositories.PatientRepository;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.web.servlet.MockMvc;
+
 import java.time.LocalDate;
 import java.util.Optional;
 import java.util.UUID;
 
-import io.undertree.symptom.domain.GivenName;
-import io.undertree.symptom.domain.TestPatientBuilder;
-import io.undertree.symptom.repositories.PatientRepository;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.web.servlet.MockMvc;
-
 import static org.hamcrest.Matchers.is;
 import static org.mockito.BDDMockito.given;
-import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-@RunWith(SpringRunner.class)
 @WebMvcTest(PatientController.class)
 public class PatientControllerTests {
 
@@ -57,9 +53,9 @@ public class PatientControllerTests {
 						.build()));
 
 		mockMvc.perform(get("/patients/e7a47ecd-4182-4209-911b-f7574ded1611")
-				.accept(APPLICATION_JSON_UTF8))
+				.accept(APPLICATION_JSON_VALUE))
 				.andExpect(status().isOk())
-				.andExpect(content().contentType(APPLICATION_JSON_UTF8))
+				.andExpect(content().contentType(APPLICATION_JSON_VALUE))
 				.andExpect(jsonPath("$.givenName", is("Guy")))
 				.andExpect(jsonPath("$.familyName", is("Stromboli")))
 				.andExpect(jsonPath("$.birthDate", is("1942-11-21")));
